+++ resources/arguments.py
@@ -99,9 +99,6 @@ If you plan to create the USB for another machine, please select the "Change Mod
         if self.args.build:
             build.BuildOpenCore(settings.custom_model or settings.computer.real_model, settings).build_opencore()
         elif self.args.patch_sys_vol:
-            if self.args.moj_cat_accel:
-                print("- Set Mojave/Catalina root patch configuration")
-                settings.moj_cat_accel = True
             print("- Set System Volume patching")
 
             if "Library/InstallerSandboxes/" in str(settings.payload_path):